Update patch set 1
Patch Set 1:

All I mean by this is why not report the correct values for the device?  The device overlay is there for the autobrightness values.  It doesn't make any sense to standardize them when they are actually different by device and we have the ability to customize that without changing other devices.

Patch-set: 1