Update patch set 1
Patch Set 1: No score

I was discussing this on XDA with another user :)
I'll try to check if HTC in an older rom build gave us the correct file (maybe they did).
Moreover...Even if they did, we should update CPU power consumption because we have a different kernel/frequencies.
Do you know a reliable way to measure it?
Is "/sys/class/power_supply/battery/current_now" measured using power profile or does it use a different way?

If we don't include a device power profile the default one is used and the measurements are wrong however :)
If HTC didn't gave us the file maybe we could try to adapt this and throw in the values we already know :)

Patch-set: 1